add shared code to util
